Implement camera movement
Implement camera rotation with mouse movement. Camera translates with
keyboard: w, s, a, and d. To enable camera movement, toggle free camera
setting on the user interface. In free camera mode, press c to use
camera rotation with mouse movement. To not turn the viewpoint of
camera suddently, make the sync up the yaw and pitch of mouse to current
camera front axis.